requirements.txt
Flask
scikit-learn
numpy
pandas
textblob
mysql-connector-python